Complete Looping System
Game will now loop completely allowing for the game to be played infinitely.  Also made minor adjustments the "clearing" state so the cards are removed from the board individually.

Shuffling, however, does not work as intended, as the position in the ds list does not change upon shuffling. Needs workaround.